+++ Assets/_Scripts/Interfaces/IAgent.cs
@@ -4,9 +4,9 @@ using UnityEngine.Events;
 public interface IAgent
 {
     int Damage { get; }
-    int Health { get; }
+    float Health { get; }
     UnityEvent OnDie { get; set; }
     UnityEvent OnGetHit { get; set; }
 
-    void GetHit(int damage, GameObject damageDealer);
+    void GetHit(float damage, GameObject damageDealer);
 }
